Made some changes in biocInstallRepos() to anticipate the de-pairing of R and BioC releases, but have not yet activated new functionality nor bumped version number.
Once activated, the new functionality will use the minor part of the
BiocInstaller package version number (y in x.y.z) to determine whether
this is a devel or release version, and set repositories accordingly.

We still need getDevel() and getRelease() functions which download
the appropriate version of BiocInstaller if you don't already have
the one you want.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/BiocInstaller@63996 bc3139a8-67e5-0310-9ffc-ced21a209358